+++ src/app/models/resources/project-add.ts
@@ -23,4 +23,5 @@ export interface ProjectAdd {
   shortDescription: string;
   description?: string;
   url: string;
+  fileId?: number;
 }